// webpack---_N_E/libs/js/sapi/src/types.ts
export interface SAPIRoomThumbnail {
  cubemapGetUrl: string
  customGetUrl: string
  expiresAt: number
  persistentGetUrl: string
  transientGetUrl: string
}

export interface SAPIAppearanceCustomization {
  profileColor: string
}

type UserAvatarStyle = "CREATOR_TOOLKIT" | "READY_PLAYER_ME" | "REALISTIC_RPM_HYBRID" | "REALISTIC"

export interface ReadyPlayerMeThumbnails {
  lowerBody?: string
  upperBody?: string
}

export interface AvatarData {
  activeAvatarStyle?: UserAvatarStyle
  avatarBody?: string
  avatarID?: string
  avatarUserID?: string
  filename?: string
  readyPlayerMeThumbnails?: ReadyPlayerMeThumbnails
  readyPlayerMeUrl?: string
  shirtColorOverride?: string
  skinColor?: string
  skinColorOverride?: string
}

export type AvatarSdkDataOnboarding = Pick<AvatarData, "avatarID" | "avatarUserID"> | null

export interface SAPIParticipantProfile {
  appearanceCustomization: SAPIAppearanceCustomization
  clientPlatform: string
  displayName: string
  id: string
  playerColor: string
  profilePicURL: string
}
// ...
export interface SAPIRoom {
  active: boolean
  activeParticipants?: SAPIParticipantProfile[]
  description: string
  directlyInvitedGuests: { [key: string]: InvitedGuestUserData }
  id: string
  isInstanceable: boolean
  isPublishedToExplore: boolean
  likedSpace: boolean
  likes: number
  lobbyType: SAPILobbyType
  name: string
  organizationOwnerID?: string
  organizationOwnerName?: string
  organizations: string[]
  owner: { displayName: string; userID: string }
  ownerID?: string
  participants: any
  polymerVersion: number
  publicLink?: boolean
  restrictedPermissions: AdministratorPermission[]
  roomAdmins: string[]
  /** All thumbnails available for a space **/
  roomThumbnails: SAPIRoomThumbnail
  shareID?: string
  shareSetting: SAPIShareSetting
  slug: string
  starred: boolean
  /** The thumbnail to use when previewing a space (i.e. in the spaces list) **/
  thumbnail: string
  thumbnails: string[]
  tileSize?: SAPITileSize
  timeLastJoined?: string
  version: number
  views: number
  /** Present on Creator Toolkit spaces */
  worldID?: string
}

export enum SAPITileSize {
  Large = "large",
  None = "",
}

export enum SAPILobbyType {
  None = "",
  Org = "ORG_LOBBY",
  Private = "PRIVATE_LOBBY",
  Public = "PUBLIC_LOBBY",
}

export type SAPIShareSetting = "ORG" | "PRIVATE" | "PUBLIC_LINK" // Public Link being phased out in DEV-5172

/** The value of the enum is the value of the actual query parameter */
export enum SAPISpaceType {
  Community = "community",
  Lobby = "lobby",
  Org = "org",
  Personal = "personal",
  PublicPark = "publicPark",
  Search = "search",
}

export interface LiveswitchMediaSettings {
  maxFrameRate: number
  maxHeight: number
  maxWidth: number
  mediaType: string
}

export interface LiveswitchConfig {
  appID: string
  mediaSettings: LiveswitchMediaSettings[]
  url: string
}

/** GET /rooms/${roomId} */
export interface GetRoomArgs {
  roomId: string
}

/** Response from GET /rooms/${roomID} */
export interface GetRoomResponse {
  roomData: SAPIRoom
}
// ...
export enum ContractType {
  ERC1155 = "erc1155",
  ERC721 = "erc721",
}

export enum BlockchainIdentifier {
  Ethereum = "ethereum",
  Matic = "matic",
  Solana = "solana",
}

export interface TokenGateConfig {
  /** If `true`, allow any token within the contract.
   * ex: token gate to owners of any cryptopunk, instead of just the owner of a specific punk
   * should only be used if the contractType is erc721
   */
  anyToken: boolean
  blockchainIdentifier: BlockchainIdentifier
  contractAddress: string
  contractType: ContractType
  disabled: boolean
  purchaseLink: string
  quantity: number
  tokenID: string
  tokenName: string
}